requirements.txt
PySide6>=6.6.0
sounddevice>=0.4.6
numpy>=1.26.0
scipy>=1.11.0
google-genai>=1.0.0
pynput>=1.7.6
python-dotenv>=1.0.0
